requirements.txt
fsspec
pyworld
librosa
omegaconf
tensorboard
matplotlib
torchcrepe
transformers
ffmpeg-python
pydub
praat-parselmouth
soundfile
scikit-learn
scipy
tqdm
torch
jax
flax
optax
orbax-checkpoint